get app and server working
