+++ ui-library-click-to-call/app/.common.babelrc.js
@@ -1,0 +1,9 @@
+const plugins = [];
+plugins.push([
+  '@babel/plugin-syntax-typescript',
+  {
+    isTSX: true
+  }
+]);
+
+module.exports = { plugins };
