+++ ui-library-click-to-call/app/package.json
@@ -96,7 +96,7 @@
     "typescript": "^4.3.5",
     "url-loader": "~4.1.1",
     "webpack": "5.76.0",
-    "webpack-cli": "^4.7.2",
+    "webpack-cli": "^5.0.2",
     "webpack-dev-server": "4.8.1"
   }
 }
