+++ ui-library-click-to-call/server/package.json
@@ -4,65 +4,64 @@
     "private": true,
     "description": "This is the server side that supports all ACS sample applications",
     "scripts": {
-      "build": "npm run clean && webpack --mode=production --env production",
-      "clean": "rimraf dist",
-      "start": "ts-node ./bin/www",
-      "start:dev": "ts-node-dev ./bin/www",
-      "start:debug": "node --inspect=9229 -r ts-node/register ./bin/www",
-      "test": "jest",
-      "prettier": "prettier --no-error-on-unmatched-pattern --write --config ./.prettierrc --ignore-path=./.prettierignore \"**/*.js\" \"**/*.jsx\" \"**/*.ts\" \"**/*.tsx\"",
-      "prettier:check": "prettier --no-error-on-unmatched-pattern --check --config ./.prettierrc --ignore-path=./.prettierignore \"**/*.js\" \"**/*.jsx\" \"**/*.ts\" \"**/*.tsx\"",
-      "lint": "eslint */**/*.{ts,tsx}",
-      "lint:fix": "eslint */**/*.{ts,tsx} --fix",
-      "lint:quiet": "eslint */**/*.{ts,tsx} --quiet"
+        "build": "npm run clean && webpack --mode=production --env production",
+        "clean": "rimraf dist",
+        "start": "ts-node ./bin/www",
+        "start:dev": "ts-node-dev ./bin/www",
+        "start:debug": "node --inspect=9229 -r ts-node/register ./bin/www",
+        "test": "jest",
+        "prettier": "prettier --no-error-on-unmatched-pattern --write --config ./.prettierrc --ignore-path=./.prettierignore \"**/*.js\" \"**/*.jsx\" \"**/*.ts\" \"**/*.tsx\"",
+        "prettier:check": "prettier --no-error-on-unmatched-pattern --check --config ./.prettierrc --ignore-path=./.prettierignore \"**/*.js\" \"**/*.jsx\" \"**/*.ts\" \"**/*.tsx\"",
+        "lint": "eslint */**/*.{ts,tsx}",
+        "lint:fix": "eslint */**/*.{ts,tsx} --fix",
+        "lint:quiet": "eslint */**/*.{ts,tsx} --quiet"
     },
     "dependencies": {
-      "@azure/communication-chat": "1.2.0",
-      "@azure/communication-common": "2.0.0",
-      "@azure/communication-identity": "1.2.0",
-      "@types/cors": "^2.8.8",
-      "cookie-parser": "~1.4.4",
-      "cors": "^2.8.5",
-      "debug": "~2.6.9",
-      "eslint-plugin-header": "^3.1.0",
-      "express": "~4.17.3",
-      "http-errors": "~1.6.3",
-      "morgan": "~1.9.1",
-      "@azure/core-http": "3.0.2"
+        "@azure/communication-chat": "1.2.0",
+        "@azure/communication-common": "2.0.0",
+        "@azure/communication-identity": "1.2.0",
+        "@azure/core-http": "3.0.2",
+        "@types/cors": "^2.8.8",
+        "cookie-parser": "~1.4.4",
+        "cors": "^2.8.5",
+        "debug": "~2.6.9",
+        "eslint-plugin-header": "^3.1.0",
+        "express": "~4.17.3",
+        "http-errors": "~1.6.3",
+        "morgan": "~1.9.1"
     },
     "devDependencies": {
-      "@types/cookie-parser": "^1.4.2",
-      "@types/copy-webpack-plugin": "^6.4.0",
-      "@types/express": "^4.17.8",
-      "@types/http-errors": "^1.8.0",
-      "@types/jest": "~26.0.22",
-      "@types/morgan": "^1.9.2",
-      "@types/node": "^14.14.10",
-      "@types/supertest": "^2.0.10",
-      "@types/webpack-node-externals": "^2.5.0",
-      "@typescript-eslint/eslint-plugin": "^5.8.1",
-      "@typescript-eslint/parser": "^5.8.1",
-      "copy-webpack-plugin": "^6.4.0",
-      "eslint": "^7.7.0",
-      "eslint-config-prettier": "^6.12.0",
-      "eslint-plugin-flowtype": "^5.2.0",
-      "eslint-plugin-import": "~2.22.1",
-      "eslint-plugin-prettier": "^3.1.4",
-      "husky": "^8.0.3",
-      "jest": "26.6.0",
-      "jest-junit": "~13.0.0",
-      "prettier": "2.3.1",
-      "pretty-quick": "^3.1.0",
-      "rimraf": "^2.6.2",
-      "supertest": "^6.0.1",
-      "ts-jest": "^26.4.4",
-      "ts-loader": "^8.0.12",
-      "ts-node": "^9.1.1",
-      "ts-node-dev": "^1.0.0",
-      "typescript": "4.3.5",
-      "webpack": "5.76.0",
-      "webpack-cli": "4.7.2",
-      "webpack-node-externals": "^2.5.2"
+        "@types/cookie-parser": "^1.4.2",
+        "@types/copy-webpack-plugin": "^6.4.0",
+        "@types/express": "^4.17.8",
+        "@types/http-errors": "^1.8.0",
+        "@types/jest": "~26.0.22",
+        "@types/morgan": "^1.9.2",
+        "@types/node": "^14.18.63",
+        "@types/supertest": "^2.0.10",
+        "@types/webpack-node-externals": "^2.5.0",
+        "@typescript-eslint/eslint-plugin": "^5.8.1",
+        "@typescript-eslint/parser": "^5.8.1",
+        "copy-webpack-plugin": "^6.4.0",
+        "eslint": "^7.7.0",
+        "eslint-config-prettier": "^6.12.0",
+        "eslint-plugin-flowtype": "^5.2.0",
+        "eslint-plugin-import": "~2.22.1",
+        "eslint-plugin-prettier": "^3.1.4",
+        "husky": "^8.0.3",
+        "jest": "26.6.0",
+        "jest-junit": "~13.0.0",
+        "prettier": "2.3.1",
+        "pretty-quick": "^3.1.0",
+        "rimraf": "^2.6.2",
+        "supertest": "^6.0.1",
+        "ts-jest": "^26.4.4",
+        "ts-loader": "^8.0.12",
+        "ts-node": "^9.1.1",
+        "ts-node-dev": "^1.0.0",
+        "typescript": "4.3.5",
+        "webpack": "5.76.0",
+        "webpack-cli": "4.7.2",
+        "webpack-node-externals": "^2.5.2"
     }
-  }
-  
+}
